test pool initial mint with fuzz, fix liquidity math lib type issues

+++ contracts/test/mocks/libraries/MockLiquidityMath.sol
@@ -7,13 +7,13 @@ contract MockLiquidityMath {
     function toAmounts(
         uint128 liquidity,
         uint160 sqrtPriceX96
-    ) external view returns (uint128 amount0, uint128 amount1) {
+    ) external view returns (uint256 amount0, uint256 amount1) {
         (amount0, amount1) = LiquidityMath.toAmounts(liquidity, sqrtPriceX96);
     }
 
     function toLiquiditySqrtPriceX96(
-        uint128 reserve0,
-        uint128 reserve1
+        uint256 reserve0,
+        uint256 reserve1
     ) external view returns (uint128 liquidity, uint160 sqrtPriceX96) {
         (liquidity, sqrtPriceX96) = LiquidityMath.toLiquiditySqrtPriceX96(
             reserve0,
